Revert "MAINT: Install kim-convergence package from pypi"
This reverts commit 5e2724a4f00b8a78ae529d7befb43e4517df5260.

After consulting with the developer of the convergence package, the
version of it released on PyPI includes a backwards-incompatible API and
is not compatible with the current version of the
DislocationCoreEnergyCubic driver.  Once the driver has been updated to
work with the PyPI release, it will be installed in the KIM developer
platform.