BUG: Fix conflation of kwargs to kimitems
Two of the subcommands for kimitems, `download` and `install` internally
call the code used by the `search` subcommand.  Because we use some
shared arguments across all three of these subcommands, the args for the
former two are passed to the latter.  This was creating two kinds of
problems: (1) non-shared args might be erroneously expected by the
search subcommand when they don't exist for the other two, and (2) the
same non-shared arg name is repeated between them.  In the case of
'download', it has no --force arg and yet this was assumed by 'search'
to exist.  In the case of 'install', the --force arg was assumed (and
did) exist in the args passed to it, but value passed was only supposed
to apply to the --force option used by 'install', not the --force option
used by 'search'.  Both of these cases have been resolved by explicitly
passing force=False to the search function.