BUG: Fix file descriptor leak
When the `pipeline-run-*` command line utils are run in --verbose mode,
file descriptors are used to be able to echo the stdout/stderr to the
terminal. Here, opened file descriptors were not being closed when they
were no longer active. This would result in exhausting all allocated
file descriptors for the 'openkim' user in the container if, for
example, running `pipeline-run-tests` or
`pipeline-run-verification-checks` when a large number of tests or
verification checks exist in the container.